Cast LFP inputs to float in the filtering/Kay pipeline
filter_ripple_band and get_Kay_ripple_consensus_trace seeded their output with
np.full_like(data, np.nan), which preserves an integer input's dtype. Integer
(e.g. raw ADC) LFPs were then truncated by that buffer and the filter/envelope
arithmetic, and the squared consensus envelope could overflow int16 and yield
sqrt(negative)=NaN. Coerce the working arrays (and filtered_lfps in
_preprocess_detector_inputs) to float. Float input is unchanged, so the
detectors remain byte-identical.

+++ ripple_detection/core.py
@@ -192,7 +192,9 @@ def filter_ripple_band(data: ArrayLike, sampling_frequency: float | None = None)
     filter_numerator, filter_denominator = _get_ripplefilter_kernel()
 
     # Validate data length
-    data_array = np.asarray(data)
+    # Cast to float so integer (e.g. raw ADC) input is not truncated by the
+    # NaN-preserving output buffer or the filter arithmetic.
+    data_array = np.asarray(data, dtype=float)
 
     # Check if data is multi-dimensional - handle NaN checking appropriately
     if data_array.ndim > 1:
